major changes
Updating images, adding new page called "cattialong" for service introduction, with some images into resource folder along with it. Also fixing the error of not being able to selected for Booking. modify the header.jsp and the Home.jsp to lead user to cattialong servlet instead of not doing anything